# Optimize Weak.find_aux and Weak.merge