--- a/Changes
+++ b/Changes
@@ -126,6 +126,9 @@ Working version
   table
   (Vincent Laviron, review by Gabriel Scherer and Daniel Bünzli)
 
+- #13740: Improve performance of Weak.find_aux
+  (Josh Berdine, review by ???)
+
 ### Other libraries:
 
 * #13376: Allow Dynlink.loadfile_private to load bytecode libraries with