--- a/stdlib/weak.ml
+++ b/stdlib/weak.ml
@@ -270,21 +270,27 @@ module Make (H : Hashtbl.HashedType) : (S with type data = H.t) = struct
   (* General auxiliary function for searching for a particular value
    * in a hash-set, and acting according to whether or not it's found *)
 
-  let find_aux t d found notfound =
+  let find_aux t d k_found k_notfound =
     let h = H.hash d in
     let index = get_index t h in
     let bucket = t.table.(index) in
     let hashes = t.hashes.(index) in
     let sz = length bucket in
-    let rec loop i =
-      if i >= sz then notfound h index
-      else if h = hashes.(i) then begin
-        match get bucket i with
-        | Some v as opt when H.equal v d -> found bucket i opt v
-        | _ -> loop (i + 1)
-      end else loop (i + 1)
-    in
-    loop 0
+    let found = ref None in
+    let i = ref 0 in
+    while !i < sz && Option.is_none !found do
+      if h = hashes.(!i) then begin
+        match get bucket !i with
+        | Some v as opt ->
+           if H.equal v d then
+             found := opt
+           else incr i
+        | _ -> incr i
+      end else incr i
+    done;
+    match !found with
+    | Some v as opt -> k_found bucket !i opt v
+    | None -> k_notfound h index
 
   let find_opt t d = find_aux t d (fun _b _i  o _v -> o)
                                   (fun _h _i -> None)