@gasche 2025-01-16
Tongue in cheek: do these changes actually matter if they don't reduce computation time? For #13737 I had the same question, but the change was very small and in fact the final code was better than the starting code (more readable), so this was easier to justify the change. Here the specialization commit introduces code duplication, so I find it much less tempting.