@jberdine 2025-01-16
Yes, it may not in fact actually matter. On the other hand, a 20% difference in allocated words for a not totally micro benchmark is not nothing, and increased allocation in isolated code often leads to increased cost of allocation (collection really) in realistic systems containing such code. So maybe worthwhile, but yes, code duplication can be objectionable.

My main motivation is that ocaml 4 is significantly faster than 5 for workloads that make heavy use of Weak or Ephemeron, and as far as I can tell, the root cause is not entirely understood. These workloads are also very sensitive to allocation behavior, for instance needing to recompute values if a previous computation stored in a Weak cache has been collected. So I see reducing allocations in Weak more as a removal of extra variables in order to potentially see the root causes more clearly than being themselves inherently valuable.