@jberdine 2025-01-17
Now trimmed down to only the one commit that replaces a nested recursive function with a while loop.